Prod 123

+++ app/src/main/java/com/neqabty/presentation/ui/notificationDetails/NotificationDetailsFragment.kt
@@ -69,12 +69,12 @@ class NotificationDetailsFragment : BaseFragment(), Injectable {
         notificationDetailsViewModel.errorState.observe(this, Observer { error ->
             showConnectionAlert(requireContext(), retryCallback = {
                 llSuperProgressbar.visibility = View.VISIBLE
-                notificationDetailsViewModel.getNotificationDetails(serviceId, serviceId, PreferencesHelper(requireContext()).user, notificationId.toInt())
+                notificationDetailsViewModel.getNotificationDetails(serviceId, serviceId, PreferencesHelper(requireContext()).user, notificationId)
             }, cancelCallback = {
                 navController().navigateUp()
             }, message = error?.message)
         })
-        notificationDetailsViewModel.getNotificationDetails(serviceId, serviceId, PreferencesHelper(requireContext()).user, notificationId.toInt())
+        notificationDetailsViewModel.getNotificationDetails(serviceId, serviceId, PreferencesHelper(requireContext()).user, notificationId)
     }
 
     fun initializeViews(notificationItem: NotificationUI) {
